+++ src/Command/Parameters/ParameterFactory.php
@@ -2,6 +2,9 @@
 
 namespace Startwind\Forrest\Command\Parameters;
 
+use Startwind\Forrest\Command\Parameters\Validation\Constraint\ConstraintFactory;
+use Startwind\Forrest\Command\Parameters\Validation\Constraint\IntegerConstraint;
+use Startwind\Forrest\Command\Parameters\Validation\Constraint\NotEmptyConstraint;
 use Startwind\Forrest\Enrichment\EnrichFunction\FunctionComposite;
 
 class ParameterFactory
@@ -11,6 +14,8 @@ class ParameterFactory
 
     public const TYPE_PASSWORD = 'forrest_password';
     public const FIELD_TYPE = 'type';
+
+    private const FIELD_CONSTRAINTS = 'constraints';
 
     /**
      * Create a Parameter configuration object from the given config array.
@@ -56,9 +61,29 @@ class ParameterFactory
             }
         }
 
+        if (array_key_exists(self::FIELD_CONSTRAINTS, $config)) {
+            $constraints = self::getConstraints($config[self::FIELD_CONSTRAINTS]);
+            $parameter->setConstraints($constraints);
+        }
+
         if (array_key_exists('enum', $config)) {
             $parameter->setValues($config['enum']);
         }
+    }
+
+    private static function getConstraints(array $constraintArray): array
+    {
+        $constraints = [];
+
+        foreach ($constraintArray as $constraint) {
+            try {
+                $constraints[] = ConstraintFactory::getConstraint($constraint);
+            } catch (\Exception $exception) {
+                // @todo log error
+            }
+        }
+
+        return $constraints;
     }
 
     /**