+++ src/Command/Parameters/Validation/SuccessfulValidationResult.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Startwind\Forrest\Command\Parameters\Validation;
+
+class SuccessfulValidationResult extends ValidationResult
+{
+    public function __construct()
+    {
+        parent::__construct(true);
+    }
+}